semaphore

+++ rootfs/lib/src/semaphore.c
@@ -11,12 +11,12 @@ int32_t semaphore_close(semaphore_t* s) {
 }
 
 int32_t semaphore_lock(semaphore_t* s) {
-	while(syscall1(SYSCALL_SEMAPHORE_LOCK, s) < 0)
+	while(syscall1(SYSCALL_SEMAPHORE_LOCK, (int32_t)s) < 0)
 		yield();
 	return 0;
 }
 
 int32_t semaphore_unlock(semaphore_t* s) {
-	return syscall1(SYSCALL_SEMAPHORE_UNLOCK, s);
+	return syscall1(SYSCALL_SEMAPHORE_UNLOCK, (int32_t)s);
 }
